Rename inheritance files to match class names.
